feature: all elements of header moved to Header folder
